Fixed residual time calculations and stuff
Residual time is calculated with the pressure group after the surface
interval, was previously being calculated with the pressure group before
that particular dive. Also noticed that user could not select a depth of
35 when adding a new dive, so I added a placeholder for the drop down.
Added more details and html to the graph's tooltips.

Currently, the pressure group for post dive is using the user-submitted
time, and not adding residual time again. Will fix shortly.